Cleaned up concept resolution
There were different subclasses of IConceptReference and
deserializers could be registered for them.
Then concepts could be resolved inside an IArea.
This all was very confusing and inconsistent.

Now a concept just has a UID string that is used for
serialization.
ILanguageRepository can be implemented and registered to
resolve a concept for a UID. That's it.