ensure cached cells are not modified
Reusing values during incremental computation is only valid
if it is guaranteed that the value doesn't change later.
All modifications must happen while the value is evaluated
and dependencies are recorded.

In MPS this is an issue that child cells are modified by their
parent editors. Preventing this from the beginning will enforce
a proper implementation.