ENTER to insert a new node into a list
In MPS, pressing ENTER actually inserts a new node
into the model, which is often even an abstract
concept or interface, so it instantiates a concept
that isn't allowed to be instantiated.

A cleaner solution is to keep this insertion as
a state in the editor and render a placeholder
into the list. Only after executing a substitute
action on that placeholder a new node is inserted
into the model.