build: always declare all publishing repos in gradle
Instead of conditionally defining maven target repositories depending on
the availability of credentials, always declare them. Users can still
select to which repository they want to publish by means of the
automatically created gradle tasks using the Maven repository names such
as publishFooToGithubPackagesRepository.

This changes makes it easier to figure out configuration problems for
the passed in credentials. Instead of simply skipping publishing because
the repo was not created on demand, now an explicit authentication
failure will be reported.