fix(model-server-lib): don't throw exception for missing reference links
reference(...) didn't check before if the reference exists on the concept and children(...) still
doesn't.
Without these checks changes in the meta-model can result in hard to debug issues, but since the
ModelQL will be replaced soon with a type-safe alternative, restoring the old behavior is just a
temporary fix.

https://issues.modelix.org/issue/MODELIX-435