feat(mps-model-adapters): re-implementation of the model-api adapters for MPS
There is an implementation of the model-api in MPS-extensions written
inside an MPS solution in baseLanguage.
Having this functionality as a separate library inside this repository
makes it much easier to maintain, update and reuse.
It also solves version conflicts, because an MPS module can have its
own version of the library loaded by its own classloader.