feat(mps-model-server-plugin): new IDEA plugin for running the LightModelServer
Currently, it's implemented in
https://github.com/modelix/modelix/tree/mps/2020.3/mps/org.modelix.model.server.mpsplugin
which publishes a separate version for each MPS version and has
dependencies on MPS-extensions.

The new plugin doesn't have any dependencies on MPS modules and there is
only one version of the plugin for all MPS versions. This makes it much
easier and faster to deploy changes.

This is also one step towards the goal of MODELIX-177 to split the
modelix/modelix repository into multiple repositories that don't have
MPS version specific releases.