feat(modelql): new query language for working with models without replication
When working with large models there is the problem that you cannot or
don't want to replicate the whole model into the client.

All the existing approaches of dealing with partially replicated models
are designed for a real-time collaboration use case.
But there are use cases that are more similar to a database application
where you render an HTML page with a summary of the data stored on the
model server.

The new ModelQL API is a query language similar to what database systems
provide. It's a powerful way of describing the data you are interested
in and then receiving just that. There is no active connection that
listens for changes and no caching required. It reduces the complexity
and solves some of the performance issues for these use cases.