fix(model-api-gen-gradle): delete caches before exporting the metamodel
Successive executions of the `export-languages` target resulted in outdated results.

The `runMPS` task in `export-languages` created an MPS cache at `{build.mps.system.path}/cache`
on the first run and reused it for the subsequent runs.

The current solution deletes the directories created by a previous run of `export-languages`
before running `export-languages` again.