fix(modelql): removed duplicate definitions of the semantic in IProducingStep
In IProducing step there were the methods createFlow, createSequence and
evaluate which all did basically the same computation, but on different
types of input.

In some cases this can have a performance benefit, but it's hard to
maintain and to guarantee that they all do the same. It's not worth to
keep them.