fix(model-api-gen): indentation in generated TypeScript files
The generator uses string templates to generate the TypeScript code.
It's hard to get the indentation right with them.
There is now a post-processing step that searches for curly brackets to
determine the correct indentation level.
Also, all the empty lines are removed, because it was inconsistent where
they were inserted.