feat(model-client): added IModelClientV2.loadVersion(RepositoryId ...)
Reading a version should always be done in the context of a repository
so that we can check the permissions.
If the version hash is leaked somewhere it still shouldn't be possible
to access the content. The permission check is not implemented yet, but
introducing the method now makes it easier to enfore the permission in
the future.