feat(bulk-model-sync-gradle): add continueOnError option
With the new option, the sync can ignore errors encountered during
syncing instead of immediately terminating. This is some sort of
graceful degradation at the cost of potentially resulting in an
inconsistent sync.

continueOnError will be implemented for local targets later on. To
inform users about this, a warning message is printed if continueOnError
is used with a local target.