fix(bulk-model-sync-lib): use carriage returns only on TTYs
The implemented progress reporting in the ModelImporter always used
carriage returns to report progress. This only works in terminals and
breaks incremental logging in setups like k8s.

This commit tries to detect if running in a terminal and if not, resorts
to logging calls for progress reporting. Logging output is reduced to
avoid excessive spam in captured log file.