feat(mps-legacy-sync-plugin): extracted plugin from MPS into new library
To improve the maintainability of the synchronization plugin between MPS
and the model-server, it was extracted from the MPS project into a new
Java/Kotlin library.

We will publish a single version of the plugin that is compatible with
a wide range of MPS version. The automated test use the intellij and
ktor test infrastructures to make them faster and more stable.

Maintaining the plugin as part of modelix.core makes it much easier to
adjust it to changes in the model-server/-client/-api.