feat(model-datastructure): new BulkUpdateOp to merge large updates into a single operation
A bulk import usually generates a lot of change operations, but an operation should capture the
intended change as high level as possible. In case of a bulk import the intention is to put the
model into the resulting state, which can be captured more accurately and more memory efficient with
the new operation.