fix(model-server): use one large transaction when putting many values
So far, when putting multiple values via objects/put, a database
transaction was created and committed for every value. That is
unbelievably slow with a real database. This commit changes the
implementation of the endpoint to use `putAll` instead of multiple `put`
calls, effectively reducing the whole operation to a single transaction.

This is a `fix` commit and not a `perf` one as the previous
implementation completely broke some known use cases due to ending in
timeouts.