perf(model-server): transfer model objects as plain text instead of JSON
VersionDelta is basically just a map of objects and their SHA hashes.
This can be serialized more efficiently with just a stream of strings separated by \n.
Using flows/sequences/streams is more memory efficient, because the data can directly be streamed
to the client without having to create this large JSON object. This should also fix the timeout
issues, because the server already starts responding while still iterating the over the model.