perf(model-server): optimized memory consumption of computeDelta
The VersionDelta computation was cached in case there are multiple clients listening for changes
on the same branch. For large updates this can consume a lot of memory. Using SoftReferences allows
the garbage collector to clear cache entries.

There was also an unnecessary caching of deltas without a second version, which resulted in caching
the whole model.