fix(build): ignore 403s when publishing and just continue
The most common publishing failure we currently see is a 403 where the
target repo indicates that the artifact in question is already present
in the repo. We don't know how that happens, but ignoring the error in
such a case is perfectly fine as the artifact is actually already
present.

This commit adds the `--continue` Gradle flag to the publish call to
ignore the 403s. To avoid trying to upload something that's completely
broken but would pass the build due to `--continue`, the `build` call
has been put into a separate workflow step and therefore Gradle

skip-lint