feat(model-server): add an option to populate caches on start
Adds a command line option for controlling whether the ignite cache is
populated eagerly on model-server start. When enabled, the model-server
only becomes responsive once the cache is populated from the database.

This option is implemented as populating the cache this way is a lot
faster than pulling a branch at the moment. For instance, in one
project, populating the cache using the loadCache method takes only a
about 20 seconds whereas pulling the data using the REST API as a means
of warming up the cache takes about 13 minutes.