fix(model-client): use streaming for large responses
According to what is discussed in
https://youtrack.jetbrains.com/issue/KTOR-5286/, use prepareGet instead
of Get for all GET requests from model-client that potentially stream
the whole model to avoid running into memory issues and
NegativeArraySize exceptions.

I was not able to come up with a working tests for this while avoiding
to actually use as much memory as MAX_INT bytes.