fix(mps-sync-plugin-lib): several fixes, see commit message body
Use a default object instead of null as the default result a CompletableFuture (uncompleted CF and CF with null result cannot be differentiated from each other). Fix FuturesWaitQueue, i.e. if a predecessor CF's result is a CF, then reenqueue it because in most cases we are curious about the CF's final result and not the CF itself. With other words: at the moment, CFs cannot be passed between SyncTasks via the last return statement, because otherwise we cannot differentiate between the case that we are waiting for an Iterable.waitForCompletionOfEach and any other custom case where a CF is the last returned object. In most cases this construct means that the result of the computation will appear at some point in the future and we want to wait for not. Rarely it means that we can pass this CF on to the next task as an input (such case is NOT possible with this implementation).