perf(bulk-model-sync): use lightweight node reference instead of nodes in critical indices
An example are `MPSNode`s and `MPSModelReference`s used during importing into MPS.
`MPSNode` (an `INode`) uses more memory because it references an `SNode` and prevents it from being garbage collected.
`MPSModelReference` (an `INodeReference`) is much lighter because it only references `SModelReference`.
`MPSModelReference` only needs to be resolved to `MPSNode` in the special case when a node was moved.