fix(model-server): InMemoryModel was blocking the ktor request handler thread pool
The model was loaded by a synchronized method and all HTTP requests were
blocking the thread while waiting for the result. Since the default pool
is constrained in size it was running out of thread for handling other
request such as the /health endpoint.

Replaced the synchronized method by coroutine features that don't block
the threads.