fix(model-api-gen): generate ITypedChildLink types instead of GeneratedChildLink
GeneratedChildLink implements ITypedChildLink and IChildLink which can
result in conflicting overloads when there are extension methods for
both interfaces.

For example `C_ClassConcept.members` was of type GeneratedChildListLink
and is now of type ITypedChildLinkList.

The same change is done for reference links and properties.

BREAKING CHANGE: The generated links/properties are not instances of
their untyped interfaces (IProperty/IChildLink/IReferenceLink) anymore.
New compile errors may appear in existing code which can be fixed by
appending `.untyped()`.