build: check binary compatibility of Kotlin JVM APIs
It's not always obvious if a change in the Kotlin code is a breaking
change of the JVM API. Sometimes annotations such as `@JvmOverloads` are
required to maintain binary compatibility.

Some changes are not breaking changes in the sense that they don't
require changes in Kotlin code that uses the API, but if the JVM API
is incompatible it can cause LinkageErrors at runtime, if the Kotlin
code wasn't re-compiled against the new version.