fix(model-datastructure): UndoOp reverted unrelated changes
If a branch did some changes and then directly reverted them, you would
expect that merging it has no effect on the resulting model, but it
actually reverted conflicting changes from the other branch.

There was already a test for that, but the number of changes it applied
was to small to create such conflicts.