fix(bulk-model-sync): update previously unresolvable references if they can be resolved
For unresolvable references we store the serialized expected reference, as we assume the node is not part of the model but the reference might still be useful for dynamic resolution on access.
If the target node became part of the model later on, we did not update it because the originalIds still matched.
Now we check, if it was previously unresolvable and update the reference to the INode accordingly.