test(model-datastructure): drop long-running test
The `knownPerformanceIssue` test case was now constantly timing out when
being executed in JS. After clarification, all corner cases it included
are already covered by the remaining tests in the test file and also by
MergeOrderTest, ConflictResolutionTest, and UndoTest. Therefore, I have
dropped this test case to avoid the timeout and to reduce the amount of
hard to maintain test cases.

The expensive part about the test was the way assertHistoryIsCorrect was
implemented, not the production code.