feat(model-server): implement deleting branches via REST API
Adds a new DELETE endpoint to the REST API for deleting a branch from a
repository.

Garbage collection of data potentially not referenced anymore is not
handled here and will be part of a later feature implementation.

The tests for ModelReplicationServer have been refactored.
runWithTestModelServer now provides a fixture data class to the executed
block containing the objects comprising the test model-server so that
tests, for instance, have access to the used repository manager.