build: replace custom detekt rules with official ones
This commit replaces our custom AvoidNonNullAssertionOperator detekt
rule with the official rules such as UnsafeCallOnNullableType. These
rules were not active because we executed detekt without type
resolution.

The GitHub workflow is now updated to use the detekt tasks that enable
type resolution. As detekt runs with type resolution use source sets to
find file to lint for, some manual exclusions for generated content or
non-jar files in case of the MPS fileTree had to be added.