chore(mps-model-adapters): fix `writeActionOnEdt` in `MpsAdaptersTestBase`
`writeActionOnEdt` did not work as expected.
The result of `ThreadUtils.runInUIThreadAndWait` was ignored.
Because of that, tests that should have failed did not fail.
Furthermore, writes needed to be executed inside a command.

Some tests need to be adapted because they were faulty.
Some tests had to be deactivated because of bugs in the implementation.