feat(model-server): report errors as application/problem+json
This commit modifies the v2 REST API to always report errors as
application/problem+json instances following RFC 9475. We do not declare
this as a breaking change as we never guaranteed a certain error format.

The OpenAPI specification has been extended with a default error
response for all routes to avoid having to repeat the definition many
times.

The installation of the StatusPages extension had to be moved to the
model-server Main file as this extension is now used to map instances of
the new HttpException class to suitable responses.

HttpException has been introduced due to the lack of a standard HTTP
exception hierarchy in ktor. Throwing exceptions instead of
call.respondXXX usually makes route implementations shorter and easier
to read.