feat(model-server): isolated storage of repository data
The versions of models where previously stored in the same storage
without knowing to which repository they belong.
To allow proper permission checks there is now an additional
`repository` field in the database.

Existing repositories remain in the global storage and don't benefit
from this new security feature. A migration is not yet implemented.

BREAKING CHANGE: New repositories are not accessible via the
    deprecated `RestWebModelClient`.
    Use IModelClientV2.initRepositoryWithLegacyStorage to create a new
    repository that is also accessible via the deprecated client.
    If you directly use the REST API to create repositories then add the
    query parameter `?legacyGlobalStorage=true`.