refactor(model-server): generate abstract base classes from OpenAPI spec
This commit sets up the generation of abstract base classes from the
OpenAPI specifications so that routes can be implemented with better
type-safety and less potential for forgetting to implement declared
routes.

As a basic example, an IdsApi has been extracted into a separate API
based on the tagging in the specification. Other implementations will
follow in separate commits / PRs.