refactor(model-server): extract the V1 API into a separate spec
This makes room to adjust the base path of the v2 API to `/v2`.

Due to the way authentication is handled, it is not straight-forward to
implement the v1 API based on the generated classes.

Splitting specs has also required adjustment to the SwaggerUI serving.
The ktor plugin does not allow serving multiple specifications on the
same route (despite genrally possible in the JS SwaggerUI). Therefore,
we simply offer multiple routes for the different API versions.