fix(model-server): fix `IgniteStoreClient.getAll` not returning values for existing entries
With different field order in Ignite configuration and key type `ObjectInRepository`, `IgniteCache.getAll` returned null values because keys were serialized differently.
This only happened when `IgniteCache.getAll` was called with more than one key and the keys were not loaded into the cache.
The serialization of a key return from a read-through operation was different from the serialization of a key created from an `ObjectInRepository` object.