build(deps): bump typescript from 4.9.5 to 5.4.2 in /ts-model-api
Replace the deprecated compiler option `importsNotUsedAsValues` with linting rules.

The new suggested compiler option `verbatimModuleSyntax` does not allow writing ESM syntax in files that will emit as CommonJS.
That is what we do.
To not rewrite the code but disable the things disabled by `importsNotUsedAsValues` the new linting rules were added.

An in-depth explanation can be found at https://johnnyreilly.com/typescript-5-importsnotusedasvalues-error-eslint-consistent-type-imports