chore: moved unused LightModelServer to tests
The handler for the light model server API inside the model-server
was only used in tests of the light-model-client, but never installed
in the model-server.

Since it's deprecated and would need to extended by permission checks
I moved it to the test source set to prevent it from being accidentally
installed in the model-server.