feat(bulk-model-sync): workaround removed and outdated resolveInfo after import into MPS
In projects, loading all libraries and plugins can significant time.
This workaround enables to sync projects better that do not want to load all libraries and plugins.
With this workaround, the `name` or `resolveInfo` property of a target is used as the `resoleInfo` property on a reference even if the concept of the target could not be fully loaded.