refactor(model-server): merge OpenAPI specs before using them
This commit makes the model-server-openapi folder a module again, which
is now responsible for merging the individual OpenAPI specification
files into a single one that's then used inside the build and for the
SwaggerUI of the model-server. This has several benefits:

* The SwaggerUI configuration and usage is easier as we present a
  single, combined API with a good overview under a single URL.
* The duplicated generated of shared models such as the Problem type is
  prevented as the merging deduplicates the data type.

Implementing merging is implemented with redocly, which was the only
tool I could find for this tasks that worked. The model-server-openapi
contains a decorator plugin for redocly that, as part of the merge
process, prepends the server URL paths to routes so that URLs are
correct in the merged result specification.

The different specifications we maintain had a few duplicated response
names that caused issues when merging. These duplications were resolved
by renaming some instances of them.

To properly categorize the different APIs in the SwaggerUI for the user,
tags were added everywhere in the specifications.

I could also get rid of several required imports by generating the
source code into the same package that's also used for the API
implementation.