chore(model-server): remove implementation of `LightModelServer` from model-server code
The implementation `org.modelix.model.server.handlers.LightModelServer` was only used as a fake to test the light-model-client.
Renaming and moving it into the test code of the light-model-client helps to understand that.

The light-model-server implementation running in MPS is `org.modelix.model.server.light.LightModelServer`.

The next step would be to remove `FakeLightModelServer` and use `org.modelix.model.server.light.LightModelServer` to test the light-model-client.
See MODELIX-994