fix(bulk-model-sync-gradle): fix MPS Runner configuration for server sources with base revision
When a base revision was specified, the specified branch was not passed as a system property to the bulk sync.

When a base revision was specified without a source revision, the property value for the source revision was the string `"null"`.

To write test with TestKit we would need further adjustments to the build setup.
For this fix, I considered them out of scope.

I tried to test execution with Gradles TestKit first.
See https://docs.gradle.org/current/userguide/test_kit.html
But the Plugin tries to download the package "org.modelix:bulk-model-sync-mps" as `mpsDependencies`.
The names of the downloaded files are then used as `classPathElements` in the MPS Runner configuration.
Therefore, the plugin cannot be tested with TestKit without publishing "org.modelix:bulk-model-sync-mps" and its dependents locally or including the "modelix.core" project from within a test execution.