build: update to ESLint 9
With the update to v9, ESLint did not support reading configuration in subdirectories.
We relied on this feature when running ESLint with pre-commit.
Therefore, the configuration was unified at the top-level of the repository.
This resulted in a cleanup of linting related build code in the `ts-model-api` and `vue-model-api` directories.
The unification makes maintenance easier anyway.

The official ESLint hook is not used anymore because of the issue described in https://github.com/eslint/eslint/issues/18465
The local hook works around that issue.
It also does not require duplicating versions of NPM packages.
But it requires installing NPM dependencies before running pre-commit hooks.