chore(model-server): deduplicate authorization logic
The logic determining the verifier was a duplicate of the logic used in the additional validation.

The logic from the validation (aka `ModelixAuthorizationConfig,nullIfInvalid`) is also used elsewhere.
For example, it is used in the workspaces through the modelix-authorization lib.
So it was reused in a custom verifier.

Calling `ModelixAuthorizationConfig.nullIfInvalid` again in `validate` is redundant.
It is now only used to create a custom principal object.