fix(model-server): make specifying a key ID optional
If no static key ID is configured, then the key ID specified in the JWT is used to verify the token.
The key is looked up at the JWK URL with the key ID from the token.

This is a fix because it restores functionality that accidentally broke in version 8.14.0.