fix(model-datastructure): role IDs are now used by default
Name based persistence is considered legacy and should not be used for
new models. To avoid accidental use of this legacy persistence all
default values where changed to `true`.

BREAKING CHANGE: The default value for `useRoleIds` parameters changed
to true. If you didn't specify a value for `useRoleIds` and rely on a
name based persistence you now have to provide an explicit `false`
value.