fix(vue-model-api): do not evict ReactiveINodeJS from cache when objects manged by it are still used
The effect of this issue was that subscribed Vue elements didn't update when a change happened in Modelix data.

Switching to `shallowRef` fixes the issue, because the created `shallowRef`s have a reference to the `ReactiveINodeJS` instances.

Implementing and using the `set` functions of `customRef`s would have had the same effect. But the implementation with `shallowRef`s is cleaner, because we also do not have to save the `trigger` functions in `ReactiveINodeJS`.

See also https://vuejs.org/guide/extras/reactivity-in-depth#integration-with-external-state-systems

Fixes: MODELIX-1041