refactor: replace Reaktive with own stream API
Gives us the flexibility to use sequences, flows or Reaktive depending on the use case.

For example, consuming a Reaktive as a flow in JS doesn't work well because of the missing
backpressure. If there is no lazy loading then sequences provide a better performance.

BREAKING CHANGE: The asynchronous stream based model API switch from Reaktive to a custom one.