refactor(model-datastructure): introduction of IObjectGraph, which controls how objects are loaded
This makes it more flexible for different use cases. There can be lazy loading or a fully
instantiated in-memory graph or something mixed.
For the model client there is an implementation that keeps the model data inside a version in
memory, but tries to unload the unused history.

BREAKING CHANGE: Some of the constructors of `CLVersion` and `CLTree` were removed. There are either
  static methods as a replacements or some of the parameters need be converted to a non-deprecated
  type before passing them to the constructor.