feat(model-datastructure): large lists of child nodes are now split into smaller objects
In one of the sandboxes in MPS-extensions is a node with 100_000 child nodes. Large objects are bad
for the scalability.

One issue is that a single key-value entry can be larger than the page size in ignite and then the
cache hasn't enough free space prepared and throws an `IgniteOutOfMemoryException`.
But large objects will cause performance issues everywhere. After using the MPS IDs instead of
int64 IDs, the average objects size increased and this problem just became more severe.