fix(model-client): ReplicatedModel was throwing LazyLoadingDisabledException on JS
The JS client doesn't support blocking requests. If during real-time collaboration the client tries
to merge the history with the remote version, it may need more of the history than what was
previously downloaded.
The algorithm in ReplicatedModel was changed to just let the server do the merge and avoid
downloading additional history to the client.