fix(mps-sync-plugin): persist whether a connection should be used for a specific repository only
When a new connection is added via `IModelSyncService.addServer` a repository ID can optionally be
provided and then the connection is only used for that repository and the JWT token will usually
also only be valid for that repository.
The `modelix.xml` previously didn't make this distinction and when the state was reloaded from disk
a previously global connection turned into a repository scoped one.