fix(mps-model-adapters): module dependencies to not loaded modules weren't synchronized properly
If a module has a dependency to a module that isn't loaded inside MPS, this dependency was written
to Modelix without the reference to the target module set.
This was originally correct but after the Modelix model data structure was changed to using the
original MPS IDs also for the nodes in Modelix we are able to synchronize even broken references
by just copying the ID of the target.
Even though, when working with the Modelix model the result is the same, the reference is still
pointing nowhere, it makes a difference when synchronizing back to MPS.